test(cli): cover package trait hashing with Swift Testing
