+++ cli/Tests/TuistHasherTests/DependenciesContentHasherTests.swift
@@ -242,83 +242,6 @@ final class DependenciesContentHasherTests: TuistUnitTestCase {
             .called(2)
     }
 
-    func test_hash_whenPackageTraitsChange_returnsDifferentHash() async throws {
-        // Given
-        let package = Package.local(path: filePath1)
-        let dependency = TargetDependency.package(product: "foo", type: .runtime)
-        let firstGraphTarget = GraphTarget.test(
-            target: Target.test(dependencies: [dependency]),
-            project: Project.test(
-                packages: [package],
-                packageTraits: [.init(package: package, traits: ["FeatureA"])]
-            )
-        )
-        let secondGraphTarget = GraphTarget.test(
-            target: Target.test(dependencies: [dependency]),
-            project: Project.test(
-                packages: [package],
-                packageTraits: [.init(package: package, traits: ["FeatureB"])]
-            )
-        )
-
-        // When
-        let firstHash = try await subject.hash(
-            graphTarget: firstGraphTarget,
-            hashedTargets: hashedTargets,
-            hashedPaths: hashedPaths
-        ).hash
-        let secondHash = try await subject.hash(
-            graphTarget: secondGraphTarget,
-            hashedTargets: hashedTargets,
-            hashedPaths: hashedPaths
-        ).hash
-
-        // Then
-        XCTAssertNotEqual(firstHash, secondHash)
-    }
-
-    func test_hash_whenPackageTraitsAreReordered_returnsTheSameHash() async throws {
-        // Given
-        let firstPackage = Package.local(path: filePath1)
-        let secondPackage = Package.local(path: filePath2)
-        let dependency = TargetDependency.package(product: "foo", type: .runtime)
-        let firstGraphTarget = GraphTarget.test(
-            target: Target.test(dependencies: [dependency]),
-            project: Project.test(
-                packages: [firstPackage, secondPackage],
-                packageTraits: [
-                    .init(package: firstPackage, traits: ["FeatureA", "FeatureB"]),
-                    .init(package: secondPackage, traits: ["FeatureC"]),
-                ]
-            )
-        )
-        let secondGraphTarget = GraphTarget.test(
-            target: Target.test(dependencies: [dependency]),
-            project: Project.test(
-                packages: [firstPackage, secondPackage],
-                packageTraits: [
-                    .init(package: secondPackage, traits: ["FeatureC"]),
-                    .init(package: firstPackage, traits: ["FeatureB", "FeatureA"]),
-                ]
-            )
-        )
-
-        // When
-        let firstHash = try await subject.hash(
-            graphTarget: firstGraphTarget,
-            hashedTargets: hashedTargets,
-            hashedPaths: hashedPaths
-        ).hash
-        let secondHash = try await subject.hash(
-            graphTarget: secondGraphTarget,
-            hashedTargets: hashedTargets,
-            hashedPaths: hashedPaths
-        ).hash
-
-        // Then
-        XCTAssertEqual(firstHash, secondHash)
-    }
-
     func test_hash_whenDependencyIsOptionalSDK_callsContentHasherAsExpected() async throws {
         // Given
         let dependency = TargetDependency.sdk(name: "foo", status: .optional)
